+++ docs/Changelog.md
@@ -13,6 +13,7 @@
 * Copy, paste playlist on save, load button context menu.
 * Fade in/out when play/stop.
 * Crossfade in timing mode.
+* Crossfade when switch track.
 ---
 #### v0.3.1.0
 ##### Changes